reformat code structure, add checking for group owner (#3)

+++ OmniDB/OmniDB_app/views/connections.py
@@ -1,57 +1,32 @@
-from django.http import HttpResponse
-from django.template import loader
-from django.http import JsonResponse
-from django.core import serializers
-from django.shortcuts import redirect
 import json
-
-import sys
-
-import OmniDB_app.include.Spartacus as Spartacus
-import OmniDB_app.include.Spartacus.Database as Database
-import OmniDB_app.include.Spartacus.Utils as Utils
-import OmniDB_app.include.OmniDatabase as OmniDatabase
-from OmniDB_app.include.Session import Session
-from OmniDB import settings
-from datetime import datetime
-
-import time, os
-
-from OmniDB_app.models.main import *
+import os
+import time
 
 import paramiko
+from django.db.models import Q
+from django.http import JsonResponse
 from sshtunnel import SSHTunnelForwarder
 
-from OmniDB_app.views.memory_objects import *
+from OmniDB import settings
+from OmniDB_app.include import OmniDatabase
+from OmniDB_app.models import Technology, Connection, Group, GroupConnection
+from OmniDB_app.views.memory_objects import user_authenticated
 
-from django.db.models import Q
 
 @user_authenticated
 def get_connections(request):
-
-    #User not authenticated
-    if not request.user.is_authenticated:
-        v_return['v_error'] = True
-        v_return['v_error_id'] = 1
-        return JsonResponse(v_return)
-
-    v_return = {}
-    v_return['v_data'] = ''
-    v_return['v_error'] = False
-    v_return['v_error_id'] = -1
+    v_return = {'v_data': '', 'v_error': False, 'v_error_id': -1}
 
     json_object = json.loads(request.POST.get('data', None))
     v_tab_conn_id_list = json_object['p_conn_id_list']
 
-    v_tech_list = []
-    for tech in Technology.objects.all():
-        v_tech_list.append(tech.name)
+    v_tech_list = [tech.name for tech in Technology.objects.all()]
 
     v_connection_list = []
-    try:
-        for conn in Connection.objects.filter(Q(user=request.user) | Q(public=True)):
+    connections = Connection.objects.filter(Q(user=request.user) | Q(public=True))
 
-            conn.user.id != request.user.id
+    if connections:
+        for conn in connections:
 
             v_conn_object = {
                 'id': conn.id,
@@ -78,7 +53,7 @@ def get_connections(request):
             if conn.id in v_tab_conn_id_list:
                 v_conn_object['locked'] = True
 
-            if (conn.technology.name!='terminal'):
+            if conn.technology.name != 'terminal':
                 v_conn_object['conn_string'] = conn.conn_string
                 v_conn_object['server'] = conn.server
                 v_conn_object['port'] = conn.port
@@ -87,9 +62,6 @@ def get_connections(request):
                 v_conn_object['password'] = False if conn.password.strip() == '' else True
 
             v_connection_list.append(v_conn_object)
-    # No connections
-    except Exception as exc:
-        None
 
     v_return['v_data'] = {
         'v_conn_list': v_connection_list,
@@ -98,58 +70,41 @@ def get_connections(request):
 
     return JsonResponse(v_return)
 
+
 @user_authenticated
 def get_groups(request):
+    v_return = {'v_data': [], 'v_error': False, 'v_error_id': -1}
 
-    v_return = {}
-    v_return['v_data'] = []
-    v_return['v_error'] = False
-    v_return['v_error_id'] = -1
-
-    #Invalid session
+    # Invalid session
     if not request.session.get('omnidb_session'):
         v_return['v_error'] = True
         v_return['v_error_id'] = 1
         return JsonResponse(v_return)
 
-    v_session = request.session.get('omnidb_session')
+    groups = Group.objects.filter(user=request.user)
 
-    v_group_list = []
-
-    v_current_group_data = {
-        'id': None,
-        'name': None,
-        'conn_list': []
-    }
-
-    try:
-        for group in Group.objects.filter(user=request.user):
+    if groups:
+        for group in groups:
             v_current_group_data = {
                 'id': group.id,
-                'name':  group.name,
+                'name': group.name,
                 'conn_list': []
             }
-            for group_conn in GroupConnection.objects.filter(group=group):
-                v_current_group_data['conn_list'].append(group_conn.connection.id)
+            group_conn = GroupConnection.objects.filter(group=group)
+            if group_conn:
+                for conn in group_conn:
+                    v_current_group_data['conn_list'].append(conn.connection.id)
 
-            v_group_list.append(v_current_group_data)
-
-    # No group connections
-    except Exception as exc:
-        None
-
-    v_return['v_data'] = v_group_list
+            v_return['v_data'].append(v_current_group_data)
 
     return JsonResponse(v_return)
 
+
 @user_authenticated
 def new_group(request):
-    v_return = {}
-    v_return['v_data'] = ''
-    v_return['v_error'] = False
-    v_return['v_error_id'] = -1
+    v_return = {'v_data': '', 'v_error': False, 'v_error_id': -1}
 
-    #Invalid session
+    # Invalid session
     if not request.session.get('omnidb_session'):
         v_return['v_error'] = True
         v_return['v_error_id'] = 1
@@ -159,7 +114,7 @@ def new_group(request):
     p_name = json_object['p_name']
 
     try:
-        new_group = Group(user=request.user,name=p_name)
+        new_group = Group(user=request.user, name=p_name)
         new_group.save()
     except Exception as exc:
         v_return['v_data'] = str(exc)
@@ -168,14 +123,10 @@ def new_group(request):
 
     return JsonResponse(v_return)
 
+
 @user_authenticated
 def edit_group(request):
-    v_return = {}
-    v_return['v_data'] = ''
-    v_return['v_error'] = False
-    v_return['v_error_id'] = -1
-
-    v_session = request.session.get('omnidb_session')
+    v_return = {'v_data': '', 'v_error': False, 'v_error_id': -1}
 
     json_object = json.loads(request.POST.get('data', None))
     p_id = json_object['p_id']
@@ -183,6 +134,13 @@ def edit_group(request):
 
     try:
         group = Group.objects.get(id=p_id)
+
+        if group.user.id != request.user.id:
+            v_return['v_data'] = 'This group does not belong to you.'
+            v_return['v_error'] = True
+
+            return JsonResponse(v_return)
+
         group.name = p_name
         group.save()
     except Exception as exc:
@@ -192,20 +150,23 @@ def edit_group(request):
 
     return JsonResponse(v_return)
 
+
 @user_authenticated
 def delete_group(request):
-    v_return = {}
-    v_return['v_data'] = ''
-    v_return['v_error'] = False
-    v_return['v_error_id'] = -1
-
-    v_session = request.session.get('omnidb_session')
+    v_return = {'v_data': '', 'v_error': False, 'v_error_id': -1}
 
     json_object = json.loads(request.POST.get('data', None))
     p_id = json_object['p_id']
 
     try:
         group = Group.objects.get(id=p_id)
+
+        if group.user.id != request.user.id:
+            v_return['v_data'] = 'This group does not belong to you.'
+            v_return['v_error'] = True
+
+            return JsonResponse(v_return)
+
         group.delete()
     except Exception as exc:
         v_return['v_data'] = str(exc)
@@ -214,57 +175,54 @@ def delete_group(request):
 
     return JsonResponse(v_return)
 
+
 @user_authenticated
 def test_connection(request):
-    v_return = {}
-    v_return['v_data'] = ''
-    v_return['v_error'] = False
-    v_return['v_error_id'] = -1
-
-    #Invalid session
+    v_return = {'v_data': '', 'v_error': False, 'v_error_id': -1}
+    # Invalid session
     if not request.session.get('omnidb_session'):
         v_return['v_error'] = True
         v_return['v_error_id'] = 1
         return JsonResponse(v_return)
 
-    v_session = request.session.get('omnidb_session')
-
     json_object = json.loads(request.POST.get('data', None))
     p_type = json_object['type']
 
-    password=json_object['password'].strip()
-    ssh_password=json_object['tunnel']['password'].strip()
-    ssh_key=json_object['tunnel']['key']
+    password = json_object['password'].strip()
+    ssh_password = json_object['tunnel']['password'].strip()
+    ssh_key = json_object['tunnel']['key']
 
-    if json_object['id']!=-1:
+    if json_object['id'] != -1:
         conn = Connection.objects.get(id=json_object['id'])
-        if json_object['password'].strip()=='':
-            password=conn.password
-        if json_object['tunnel']['password'].strip()=='':
-            ssh_password=conn.ssh_password
-        if json_object['tunnel']['key'].strip()=='':
-            ssh_key=conn.ssh_key
+        if json_object['password'].strip() == '':
+            password = conn.password
+        if json_object['tunnel']['password'].strip() == '':
+            ssh_password = conn.ssh_password
+        if json_object['tunnel']['key'].strip() == '':
+            ssh_key = conn.ssh_key
 
-    if json_object['temp_password']!=None:
+    if json_object['temp_password']:
         password = json_object['temp_password']
 
-
-    if p_type=='terminal':
+    if p_type == 'terminal':
 
         client = paramiko.SSHClient()
         client.load_system_host_keys()
         client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
 
         try:
-            #ssh key provided
+            # ssh key provided
             if ssh_key.strip() != '':
-                v_file_name = '{0}'.format(str(time.time())).replace('.','_')
+                v_file_name = '{0}'.format(str(time.time())).replace('.', '_')
                 v_full_file_name = os.path.join(settings.TEMP_DIR, v_file_name)
-                with open(v_full_file_name,'w') as f:
+                with open(v_full_file_name, 'w') as f:
                     f.write(ssh_key)
-                client.connect(hostname=json_object['tunnel']['server'],username=json_object['tunnel']['user'],key_filename=v_full_file_name,passphrase=ssh_password,port=int(json_object['tunnel']['port']))
+                client.connect(hostname=json_object['tunnel']['server'], username=json_object['tunnel']['user'],
+                               key_filename=v_full_file_name, passphrase=ssh_password,
+                               port=int(json_object['tunnel']['port']))
             else:
-                client.connect(hostname=json_object['tunnel']['server'],username=json_object['tunnel']['user'],password=ssh_password,port=int(json_object['tunnel']['port']))
+                client.connect(hostname=json_object['tunnel']['server'], username=json_object['tunnel']['user'],
+                               password=ssh_password, port=int(json_object['tunnel']['port']))
 
             client.close()
             v_return['v_data'] = 'Connection successful.'
@@ -282,24 +240,24 @@ def test_connection(request):
             password,
             -1,
             '',
-            p_conn_string = json_object['connstring'],
-            p_parse_conn_string = True
+            p_conn_string=json_object['connstring'],
+            p_parse_conn_string=True
         )
 
         # create tunnel if enabled
-        if json_object['tunnel']['enabled'] == True:
+        if json_object['tunnel']['enabled']:
 
             try:
                 if ssh_key.strip() != '':
-                    v_file_name = '{0}'.format(str(time.time())).replace('.','_')
+                    v_file_name = '{0}'.format(str(time.time())).replace('.', '_')
                     v_full_file_name = os.path.join(settings.TEMP_DIR, v_file_name)
-                    with open(v_full_file_name,'w') as f:
+                    with open(v_full_file_name, 'w') as f:
                         f.write(ssh_key)
                     server = SSHTunnelForwarder(
                         (json_object['tunnel']['server'], int(json_object['tunnel']['port'])),
                         ssh_username=json_object['tunnel']['user'],
                         ssh_private_key_password=ssh_password,
-                        ssh_pkey = v_full_file_name,
+                        ssh_pkey=v_full_file_name,
                         remote_bind_address=(database.v_active_server, int(database.v_active_port)),
                         logger=None
                     )
@@ -335,18 +293,10 @@ def test_connection(request):
 
     return JsonResponse(v_return)
 
+
 @user_authenticated
 def save_connection(request):
-    v_return = {}
-    v_return['v_data'] = ''
-    v_return['v_error'] = False
-    v_return['v_error_id'] = -1
-
-    #User not authenticated
-    if not request.user.is_authenticated:
-        v_return['v_error'] = True
-        v_return['v_error_id'] = 1
-        return JsonResponse(v_return)
+    v_return = {'v_data': '', 'v_error': False, 'v_error_id': -1}
 
     v_session = request.session.get('omnidb_session')
 
@@ -375,7 +325,7 @@ def save_connection(request):
 
             )
             conn.save()
-        #update
+        # update
         else:
             conn = Connection.objects.get(id=p_id)
 
@@ -384,25 +334,25 @@ def save_connection(request):
                 v_return['v_error'] = True
                 return JsonResponse(v_return)
 
-            conn.technology=Technology.objects.get(name=json_object['type'])
-            conn.server=json_object['server']
-            conn.port=json_object['port']
-            conn.database=json_object['database']
-            conn.username=json_object['user']
-            if json_object['password'].strip()!='':
-                conn.password=json_object['password']
-            conn.alias=json_object['title']
-            conn.ssh_server=json_object['tunnel']['server']
-            conn.ssh_port=json_object['tunnel']['port']
-            conn.ssh_user=json_object['tunnel']['user']
-            if json_object['tunnel']['password'].strip()!='':
-                conn.ssh_password=json_object['tunnel']['password']
-            if json_object['tunnel']['key'].strip()!='':
-                conn.ssh_key=json_object['tunnel']['key']
+            conn.technology = Technology.objects.get(name=json_object['type'])
+            conn.server = json_object['server']
+            conn.port = json_object['port']
+            conn.database = json_object['database']
+            conn.username = json_object['user']
+            if json_object['password'].strip() != '':
+                conn.password = json_object['password']
+            conn.alias = json_object['title']
+            conn.ssh_server = json_object['tunnel']['server']
+            conn.ssh_port = json_object['tunnel']['port']
+            conn.ssh_user = json_object['tunnel']['user']
+            if json_object['tunnel']['password'].strip() != '':
+                conn.ssh_password = json_object['tunnel']['password']
+            if json_object['tunnel']['key'].strip() != '':
+                conn.ssh_key = json_object['tunnel']['key']
 
-            conn.use_tunnel=json_object['tunnel']['enabled']
-            conn.conn_string=json_object['connstring']
-            conn.public=public=json_object['public']
+            conn.use_tunnel = json_object['tunnel']['enabled']
+            conn.conn_string = json_object['connstring']
+            conn.public = json_object['public']
             conn.save()
 
         tunnel_information = {
@@ -423,15 +373,14 @@ def save_connection(request):
             conn.password,
             conn.id,
             conn.alias,
-            p_conn_string = conn.conn_string,
-            p_parse_conn_string = True
+            p_conn_string=conn.conn_string,
+            p_parse_conn_string=True
         )
 
         prompt_password = conn.password == ''
 
-        v_session.AddDatabase(conn.id,conn.technology.name,database,prompt_password,tunnel_information,conn.alias, json_object['public'])
-
-
+        v_session.AddDatabase(conn.id, conn.technology.name, database, prompt_password, tunnel_information, conn.alias,
+                              json_object['public'])
 
     except Exception as exc:
         v_return['v_data'] = str(exc)
@@ -442,18 +391,10 @@ def save_connection(request):
 
     return JsonResponse(v_return)
 
+
 @user_authenticated
 def delete_connection(request):
-    v_return = {}
-    v_return['v_data'] = ''
-    v_return['v_error'] = False
-    v_return['v_error_id'] = -1
-
-    #User not authenticated
-    if not request.user.is_authenticated:
-        v_return['v_error'] = True
-        v_return['v_error_id'] = 1
-        return JsonResponse(v_return)
+    v_return = {'v_data': '', 'v_error': False, 'v_error_id': -1}
 
     v_session = request.session.get('omnidb_session')
 
@@ -479,18 +420,10 @@ def delete_connection(request):
 
     return JsonResponse(v_return)
 
+
 @user_authenticated
 def save_group_connections(request):
-    v_return = {}
-    v_return['v_data'] = ''
-    v_return['v_error'] = False
-    v_return['v_error_id'] = -1
-
-    #User not authenticated
-    if not request.user.is_authenticated:
-        v_return['v_error'] = True
-        v_return['v_error_id'] = 1
-        return JsonResponse(v_return)
+    v_return = {'v_data': '', 'v_error': False, 'v_error_id': -1}
 
     json_object = json.loads(request.POST.get('data', None))
     p_group = json_object['p_group']
@@ -501,7 +434,8 @@ def save_group_connections(request):
     for v_conn_data in p_conn_data_list:
         try:
             if not v_conn_data['selected']:
-                conn = GroupConnection.objects.get(group=group_obj,connection=Connection.objects.get(id=v_conn_data['id']))
+                conn = GroupConnection.objects.get(group=group_obj,
+                                                   connection=Connection.objects.get(id=v_conn_data['id']))
                 conn.delete()
             else:
                 conn = GroupConnection(
